refactor: 테스트 패키지 경로 변경
