+++ src/test/java/ladder/model/ladder/LadderRowTest.java
@@ -1,6 +1,5 @@
-package ladder.model;
+package ladder.model.ladder;
 
-import ladder.model.ladder.LadderRow;
 import ladder.strategy.RandomStrategy;
 import org.junit.jupiter.api.Test;
